--- santa/src/Santa.Logic/Objects/Address.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Santa.Logic.Objects
-{
-    public class Address
-    {
-        public string addressLineOne { get; set; }
-        public string addressLineTwo { get; set; }
-        public string city { get; set; }
-        public string state { get; set; }
-        public string postalCode { get; set; }
-        public string country { get; set; }
-    }
-}
