+++ santa/src/Santa.Logic/Objects/Base Objects/AssignmentStatus.cs
