+++ santa/src/Santa.Logic/Objects/Base Objects/BoardEntry.cs
