+++ santa/src/Santa.Logic/Objects/Base Objects/EntryType.cs
