+++ santa/src/Santa.Logic/Objects/Base Objects/Message.cs
