+++ santa/src/Santa.Logic/Objects/Information Objects/AllowedAssignmentMeta.cs
