+++ santa/src/Santa.Logic/Objects/Information Objects/ClientMeta.cs
