+++ santa/src/Santa.Logic/Objects/Logic Models/MessageHistory.cs
