+++ santa/src/Santa.Logic/Objects/Logic Models/SearchQueries.cs
