--- santa/src/Santa.Logic/Objects/Status.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Santa.Logic.Objects
-{
-    public class Status
-    {
-        public Guid statusID { get; set; }
-        public string statusDescription { get; set; }
-    }
-}
